+++ api/web/src/components/CloudTAK/PluginPane.vue
@@ -32,7 +32,7 @@ const pluginItems = computed(() => {
 .plugin-pane {
     z-index: 5;
     left: 50%;
-    bottom: 8px;
+    bottom: calc(8px + var(--map-bottom-inset, 0px));
     transform: translateX(-50%);
     height: 50px;
     overflow: visible;
@@ -40,7 +40,7 @@ const pluginItems = computed(() => {
 
 @media (max-width: 600px) {
     .plugin-pane {
-        bottom: calc(var(--map-gps-panel-size, 110px) + 16px);
+        bottom: calc(var(--map-gps-panel-size, 110px) + 16px + var(--map-bottom-inset, 0px));
     }
 }
 </style>